Update remaining resources to use class based generator structure
